requirements.txt
neo4j==5.23.1
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
openai==1.64.0
pypdf==4.3.1

